refactor: investor password
