+++ packages/trader/src/Modules/CFD/Containers/cfd-password-manager-modal.jsx
@@ -1,12 +1,9 @@
-import { Field, Form, Formik } from 'formik';
 import PropTypes from 'prop-types';
 import React from 'react';
 import {
     Icon,
     Modal,
     Tabs,
-    PasswordInput,
-    PasswordMeter,
     Button,
     DesktopWrapper,
     Div100vhContainer,
@@ -29,6 +26,7 @@ import {
 import { connect } from 'Stores/connect';
 import CFDStore from 'Stores/Modules/CFD/cfd-store';
 import TradingPasswordManager from './trading-password-manager';
+import InvestorPasswordManager from './investor-password-manager.jsx';
 
 const CountdownComponent = ({ count_from = 60, onTimeout }) => {
     const [count, setCount] = React.useState(count_from);
@@ -127,130 +125,9 @@ const CFDPasswordReset = ({ sendVerifyEmail, account_type, account_group, server
     );
 };
 
-const CFDPasswordSuccessMessage = ({ toggleModal, is_investor }) => (
-    <div className='cfd-password-manager__success'>
-        <Icon icon='IcPasswordUpdated' size={128} />
-        <Text as='p' size='xxs' align='center'>
-            {is_investor ? (
-                <Localize i18n_default_text='Your investor password has been changed.' />
-            ) : (
-                <Localize i18n_default_text='Your password has been changed.' />
-            )}
-        </Text>
-        <Button onClick={toggleModal} className='cfd-password-manager__success-btn' primary large>
-            <p className='dc-btn__text'>{localize('OK')}</p>
-        </Button>
-    </div>
-);
-
 const CFDPasswordManagerTabContentWrapper = ({ multi_step_ref, steps }) => (
     <MultiStep ref={multi_step_ref} steps={steps} className='cfd-password-manager' lbl_previous={localize('Back')} />
 );
-
-const InvestorPasswordManager = ({
-    error_message_investor,
-    is_submit_success_investor,
-    multi_step_ref,
-    onSubmit,
-    setPasswordType,
-    toggleModal,
-    validatePassword,
-}) => {
-    if (is_submit_success_investor) {
-        return <CFDPasswordSuccessMessage toggleModal={toggleModal} is_investor />;
-    }
-
-    const initial_values = { old_password: '', new_password: '', password_type: 'investor' };
-
-    return (
-        <div className='cfd-password-manager__investor-wrapper'>
-            <Text as='p' size='xs' className='cfd-password-manager--paragraph'>
-                <Localize i18n_default_text='Use this password to grant viewing access to another user. While they may view your trading account, they will not be able to trade or take any other actions.' />
-            </Text>
-            <Text as='p' size='xs' className='cfd-password-manager--paragraph'>
-                <Localize i18n_default_text='If this is the first time you try to create a password, or you have forgotten your password, please reset it.' />
-            </Text>
-            {error_message_investor && (
-                <Text as='p' color='loss-danger' size='xs' className='cfd-password-manager--error-message'>
-                    {error_message_investor}
-                </Text>
-            )}
-            <Formik initialValues={initial_values} validate={validatePassword} onSubmit={onSubmit}>
-                {({ isSubmitting, errors, setFieldTouched, values, touched }) => (
-                    <Form className='cfd-password-manager__investor-form' noValidate>
-                        <Field name='old_password'>
-                            {({ field }) => (
-                                <PasswordInput
-                                    {...field}
-                                    autoComplete='current-password'
-                                    label={localize('Current investor password')}
-                                    error={touched.old_password && errors.old_password}
-                                    required
-                                />
-                            )}
-                        </Field>
-                        <Field name='new_password'>
-                            {({ field }) => (
-                                <PasswordMeter
-                                    input={field.value}
-                                    has_error={!!(touched.new_password && errors.new_password)}
-                                    custom_feedback_messages={getErrorMessages().password_warnings}
-                                >
-                                    {({ has_warning }) => (
-                                        <PasswordInput
-                                            {...field}
-                                            autoComplete='new-password'
-                                            label={localize('New investor password')}
-                                            hint={
-                                                !has_warning &&
-                                                localize(
-                                                    'Strong passwords contain at least 8 characters, combine uppercase and lowercase letters and numbers.'
-                                                )
-                                            }
-                                            error={touched.new_password && errors.new_password}
-                                            onChange={e => {
-                                                setFieldTouched('new_password', true, true);
-                                                field.onChange(e);
-                                            }}
-                                            className='cfd-password-manager__new-password'
-                                            required
-                                        />
-                                    )}
-                                </PasswordMeter>
-                            )}
-                        </Field>
-                        <div className='cfd-password-manager__actions'>
-                            <Button
-                                className='cfd-password-manager--button'
-                                is_disabled={
-                                    isSubmitting ||
-                                    !values.old_password ||
-                                    !values.new_password ||
-                                    Object.keys(errors).length > 0
-                                }
-                                is_loading={isSubmitting}
-                                text={localize('Change investor password')}
-                                primary
-                                large
-                            />
-                            <Button
-                                className='cfd-password-manager--button'
-                                type='button'
-                                onClick={() => {
-                                    setPasswordType('investor');
-                                    multi_step_ref.current?.goNextStep();
-                                }}
-                                text={localize('Create or reset investor password')}
-                                tertiary
-                                large
-                            />
-                        </div>
-                    </Form>
-                )}
-            </Formik>
-        </div>
-    );
-};
 
 const CFDPasswordManagerTabContent = ({
     toggleModal,
